requirements.txt
numpy==1.23.5
Pillow==9.5.0
#torch==2.1.0+cu118
#torchvision==0.16.0+cu118
tqdm==4.66.3
mlflow
